refactor(core): invert imports
