+++ packages/account/user-login-widget/login/user-login.model.ts
@@ -1,7 +1,1 @@
-export interface UserLogin {
-    id?: any;
-    typeKey?: string;
-    stateKey?: string;
-    login?: string;
-    removed?: boolean;
-}
+export { UserLogin } from '@xm-ngx/core/user';
